refacto : create a rule component to store all business rules
